backend/requirements.txt
# RoadVisionAI — Docker / Linux requirements (CPU inference)
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
uvloop>=0.19.0
httptools>=0.6.0
websockets>=12.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
python-multipart>=0.0.9
onnxruntime>=1.17.3
httpx>=0.28.1